+++ src/test/java/chess/dao/WebBoardDaoTest.java
@@ -3,7 +3,7 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.game.ChessBoard;
+import chess.domain.game.ChessGame;
 import chess.domain.pieces.Color;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
@@ -22,20 +22,20 @@ class WebBoardDaoTest {
 
     @BeforeEach
     void setup() {
-        final ChessBoard board = dao.save(new ChessBoard("개초보만", "1111"));
+        final ChessGame board = dao.save(new ChessGame("개초보만", "1111"));
         this.boardId = board.getId();
     }
 
     @Test
     void saveTest() {
-        final ChessBoard board = dao.save(new ChessBoard("개초보만", "1111"));
+        final ChessGame board = dao.save(new ChessGame("개초보만", "1111"));
         assertThat(board.getRoomTitle()).isEqualTo("개초보만");
     }
 
     @Test
     void getByIdTest() {
-        final ChessBoard board = dao.save(new ChessBoard("개초보만", Color.WHITE, "2222"));
-        final ChessBoard foundBoard = dao.getById(board.getId());
+        final ChessGame board = dao.save(new ChessGame("개초보만", Color.WHITE, "2222"));
+        final ChessGame foundBoard = dao.getById(board.getId());
         assertAll(
                 () -> assertThat(foundBoard.getRoomTitle()).isEqualTo("개초보만"),
                 () -> assertThat(foundBoard.getTurn()).isEqualTo(Color.WHITE),
@@ -45,7 +45,7 @@ class WebBoardDaoTest {
 
     @Test
     void deleteBoard() {
-        final ChessBoard board = dao.save(new ChessBoard("aaa", "1111"));
+        final ChessGame board = dao.save(new ChessGame("aaa", "1111"));
         int affectedRow = dao.deleteByIdAndPassword(board.getId(), board.getPassword());
         assertThat(affectedRow).isEqualTo(1);
     }
@@ -58,7 +58,7 @@ class WebBoardDaoTest {
     @Test
     void deleteAll() {
         dao.deleteAll();
-        List<ChessBoard> boards = dao.findAll();
+        List<ChessGame> boards = dao.findAll();
 
         assertThat(boards.size()).isEqualTo(0);
     }
@@ -66,7 +66,7 @@ class WebBoardDaoTest {
     @Test
     void updateTurn() {
         dao.updateTurn(Color.BLACK, boardId);
-        ChessBoard chessBoard = dao.getById(boardId);
+        ChessGame chessBoard = dao.getById(boardId);
 
         assertThat(chessBoard.getTurn()).isEqualTo(Color.BLACK);
     }
